vault-droid/requirements.txt
kubernetes
hvac
pyyaml
cryptography